+++ src/NAudio.Core/Wave/WaveOutputs/WaveFileWriter.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using NAudio.Wave.SampleProviders;
 using NAudio.Utils;
+using NAudio.Dmo;
 
 // ReSharper disable once CheckNamespace
 namespace NAudio.Wave;
@@ -405,7 +406,18 @@ public class WaveFileWriter : Stream
         }
         else if (WaveFormat.BitsPerSample == 32 && WaveFormat.Encoding == WaveFormatEncoding.Extensible)
         {
-            writer.Write((Int32)(Int32.MaxValue * sample));
+            // A 32-bit WAVE_FORMAT_EXTENSIBLE can be either integer PCM or IEEE float,
+            // distinguished by its SubFormat GUID (NAudio's WaveFormatExtensible defaults
+            // 32-bit to IEEE float). Honour the declared subformat rather than assuming PCM.
+            if (WaveFormat is WaveFormatExtensible { SubFormat: var subFormat } &&
+                subFormat == AudioMediaSubtypes.MEDIASUBTYPE_IEEE_FLOAT)
+            {
+                writer.Write(sample);
+            }
+            else
+            {
+                writer.Write((Int32)(Int32.MaxValue * sample));
+            }
             dataChunkSize += 4;
         }
         else if (WaveFormat.Encoding == WaveFormatEncoding.IeeeFloat)